+++ app/src/main/java/io/github/aoemerson/crimesmvp/model/PoliceRestClient.java
@@ -2,6 +2,7 @@ package io.github.aoemerson.crimesmvp.model;
 
 import java.util.List;
 
+import io.github.aoemerson.crimesmvp.model.data.Crime;
 import retrofit2.Call;
 import retrofit2.http.GET;
 import retrofit2.http.Query;
